# client/src/types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Eyened API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/instances/{instance_id}": {
      "get": {
        "summary": "Get Instance",
        "description": "Get a single image instance by id, with optional related graphs.",
        "operationId": "get_instance_instances__instance_id__get",
        "parameters": [
          {
            "name": "instance_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Instance Id"
            }
          },
          {
            "name": "with_segmentations",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "With Segmentations"
            }
          },
          {
            "name": "with_form_annotations",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "With Form Annotations"
            }
          },
          {
            "name": "with_model_segmentations",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "With Model Segmentations"
            }
          },
          {
            "name": "with_tag_metadata",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "With Tag Metadata"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          },
          {
            "name": "jwt_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Jwt Token"
            }
          },
          {
            "name": "refresh_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Refresh Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageGET"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/images/{image_id}": {
      "get": {
        "summary": "Get Public Image",
        "description": "Get a single image instance by PublicID, with optional related graphs.",
        "operationId": "get_public_image_images__image_id__get",
        "parameters": [
          {
            "name": "image_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Image Id"
            }
          },
          {
            "name": "with_segmentations",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "With Segmentations"
            }
          },
          {
            "name": "with_form_annotations",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "With Form Annotations"
            }
          },
          {
            "name": "with_model_segmentations",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "With Model Segmentations"
            }
          },
          {
            "name": "with_tag_metadata",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "With Tag Metadata"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          },
          {
            "name": "jwt_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Jwt Token"
            }
          },
          {
            "name": "refresh_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Refresh Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageGET"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/images/{image_id}/data": {
      "get": {
        "summary": "Get Public Image Data",
        "description": "Redirect to the stored image data for an instance (by PublicID).",
        "operationId": "get_public_image_data_images__image_id__data_get",
        "parameters": [
          {
            "name": "image_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Image Id"
            }
          },
          {
            "name": "index",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Index"
            }
          },
          {
            "name": "meta",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Meta"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          },
          {
            "name": "jwt_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Jwt Token"
            }
          },
          {
            "name": "refresh_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Refresh Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/images/{image_id}/thumbnail": {
      "get": {
        "summary": "Get Public Image Thumbnail",
        "description": "Redirect to the stored thumbnail for an instance (by PublicID).",
        "operationId": "get_public_image_thumbnail_images__image_id__thumbnail_get",
        "parameters": [
          {
            "name": "image_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Image Id"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 144,
              "title": "Size"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          },
          {
            "name": "jwt_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Jwt Token"
            }
          },
          {
            "name": "refresh_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Refresh Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/instances/{instance_id}/tags": {
      "post": {
        "summary": "Tag Instance",
        "description": "Attach a Tag to an ImageInstance by tag ID (idempotent).",
        "operationId": "tag_instance_instances__instance_id__tags_post",
        "parameters": [
          {
            "name": "instance_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Instance Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          },
          {
            "name": "jwt_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Jwt Token"
            }
          },
          {
            "name": "refresh_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Refresh Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObjectTagPOST"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagMeta"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/instances/{instance_id}/tags/{tag_id}": {
      "patch": {
        "summary": "Patch Instance Tag",
        "description": "Update comment on an existing ImageInstance tag link.",
        "operationId": "patch_instance_tag_instances__instance_id__tags__tag_id__patch",
        "parameters": [
          {
            "name": "instance_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Instance Id"
            }
          },
          {
            "name": "tag_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tag Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          },
          {
            "name": "jwt_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Jwt Token"
            }
          },
          {
            "name": "refresh_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Refresh Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObjectTagPATCH"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagMeta"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Untag Instance",
        "description": "Remove a Tag from an ImageInstance (idempotent).",
        "operationId": "untag_instance_instances__instance_id__tags__tag_id__delete",
        "parameters": [
          {
            "name": "instance_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Instance Id"
            }
          },
          {
            "name": "tag_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tag Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          },
          {
            "name": "jwt_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Jwt Token"
            }
          },
          {
            "name": "refresh_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Refresh Token"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreatorMeta": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "CreatorMeta"
      },
...
      "DeviceMeta": {
        "properties": {
          "manufacturer": {
            "type": "string",
            "title": "Manufacturer"
          },
          "model": {
            "type": "string",
            "title": "Model"
          }
        },
        "type": "object",
        "required": [
          "manufacturer",
          "model"
        ],
        "title": "DeviceMeta"
      },
...
      "ETDRSField": {
        "type": "string",
        "enum": [
          "F1",
          "F2",
          "F3",
          "F4",
          "F5",
          "F6",
          "F7"
        ],
        "title": "ETDRSField"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImageGET": {
        "properties": {
          "sop_instance_uid": {
            "type": "string",
            "title": "Sop Instance Uid"
          },
          "data_format": {
            "type": "string",
            "title": "Data Format"
          },
          "data_source_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Source Id"
          },
          "modality": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Modality"
              },
              {
                "type": "null"
              }
            ]
          },
          "dicom_modality": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModalityType"
              },
              {
                "type": "null"
              }
            ]
          },
          "etdrs_field": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ETDRSField"
              },
              {
                "type": "null"
              }
            ]
          },
          "angio_graphy": {
            "type": "string",
            "title": "Angio Graphy"
          },
          "laterality": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Laterality"
              },
              {
                "type": "null"
              }
            ]
          },
          "anatomic_region": {
            "type": "string",
            "title": "Anatomic Region"
          },
          "rows": {
            "type": "integer",
            "title": "Rows"
          },
          "columns": {
            "type": "integer",
            "title": "Columns"
          },
          "nr_of_frames": {
            "type": "integer",
            "title": "Nr Of Frames"
          },
          "resolution_horizontal": {
            "type": "number",
            "title": "Resolution Horizontal"
          },
          "resolution_vertical": {
            "type": "number",
            "title": "Resolution Vertical"
          },
          "resolution_axial": {
            "type": "number",
            "title": "Resolution Axial"
          },
          "cf_roi": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cf Roi"
          },
          "cf_keypoints": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cf Keypoints"
          },
          "cf_quality": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cf Quality"
          },
          "date_inserted": {
            "type": "string",
            "format": "date-time",
            "title": "Date Inserted"
          },
          "date_modified": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Modified"
          },
          "date_preprocessed": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Preprocessed"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project": {
            "$ref": "#/components/schemas/ProjectMeta"
          },
          "patient": {
            "$ref": "#/components/schemas/PatientMeta"
          },
          "study": {
            "$ref": "#/components/schemas/StudyMeta"
          },
          "series": {
            "$ref": "#/components/schemas/SeriesMeta"
          },
          "device": {
            "$ref": "#/components/schemas/DeviceMeta"
          },
          "scan": {
            "$ref": "#/components/schemas/ScanMeta"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/TagMeta"
            },
            "type": "array",
            "title": "Tags"
          },
          "segmentations": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SegmentationGET"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Segmentations"
          },
          "model_segmentations": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ModelSegmentationGET"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Segmentations"
          },
          "form_annotations": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FormAnnotationGET"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Form Annotations"
          },
          "model_attrs": {
            "additionalProperties": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "object",
            "title": "Model Attrs"
          },
          "attrs": {
            "additionalProperties": true,
            "type": "object",
            "title": "Attrs"
          }
        },
        "type": "object",
        "required": [
          "sop_instance_uid",
          "data_format",
          "angio_graphy",
          "anatomic_region",
          "rows",
          "columns",
          "nr_of_frames",
          "resolution_horizontal",
          "resolution_vertical",
          "resolution_axial",
          "date_inserted",
          "id",
          "project",
          "patient",
          "study",
          "series",
          "device",
          "scan",
          "tags",
          "model_attrs",
          "attrs"
        ],
        "title": "ImageGET"
      },
...
      "Laterality": {
        "type": "string",
        "enum": [
          "L",
          "R"
        ],
        "title": "Laterality"
      },
      "Modality": {
        "type": "string",
        "enum": [
          "AdaptiveOptics",
          "ColorFundus",
          "ColorFundusStereo",
          "RedFreeFundus",
          "ExternalEye",
          "LensPhotograph",
          "Ophthalmoscope",
          "Autofluorescence",
          "FluoresceinAngiography",
          "ICGA",
          "InfraredReflectance",
          "BlueReflectance",
          "GreenReflectance",
          "OCT",
          "OCTA"
        ],
        "title": "Modality"
      },
      "ModalityType": {
        "type": "string",
        "enum": [
          "OP",
          "OPT",
          "SC"
        ],
        "title": "ModalityType"
      },
...
      "ObjectTagPATCH": {
        "properties": {
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "title": "ObjectTagPATCH"
      },
      "ObjectTagPOST": {
        "properties": {
          "tag_id": {
            "type": "integer",
            "title": "Tag Id"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "required": [
          "tag_id"
        ],
        "title": "ObjectTagPOST"
      },
...
      "PatientMeta": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "identifier": {
            "type": "string",
            "title": "Identifier"
          },
          "birth_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Date"
          },
          "sex": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SexEnum"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "identifier"
        ],
        "title": "PatientMeta"
      },
      "ProjectMeta": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "ProjectMeta"
      },
...
      "ScanMeta": {
        "properties": {
          "mode": {
            "type": "string",
            "title": "Mode"
          }
        },
        "type": "object",
        "required": [
          "mode"
        ],
        "title": "ScanMeta"
      },
...
      "SeriesMeta": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "SeriesMeta"
      },
      "SexEnum": {
        "type": "string",
        "enum": [
          "M",
          "F"
        ],
        "title": "SexEnum"
      },
...
      "StudyMeta": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "date": {
            "type": "string",
            "title": "Date"
          }
        },
        "type": "object",
        "required": [
          "id",
          "date"
        ],
        "title": "StudyMeta"
      },
...
      "TagMeta": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "tagger": {
            "$ref": "#/components/schemas/CreatorMeta"
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "title": "Date"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "tagger",
          "date"
        ],
        "title": "TagMeta"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}